Add blank activity for sync
