+++ mobile/build.gradle
@@ -76,6 +76,7 @@ dependencies {
 
     implementation 'androidx.activity:activity-compose:1.7.2'
     implementation 'com.google.android.gms:play-services-ads-lite:22.2.0'
+    implementation 'androidx.work:work-runtime-ktx:2.8.1'
     def compose_ui_version = captureVersion(implementation('androidx.compose.ui:ui:1.4.3'))
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_ui_version"
     implementation "androidx.compose.ui:ui-viewbinding:$compose_ui_version"
