+++ mobile/src/main/AndroidManifest.xml
@@ -22,6 +22,19 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name=".sync.SyncActivity"
+            android:exported="true"
+            android:theme="@android:style/Theme.Translucent.NoTitleBar"
+            android:launchMode="singleTop">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+                <data android:scheme="httprequesttile" android:host="sync" />
+            </intent-filter>
+        </activity>
+
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
             android:value="${admob_key}"/>
