+++ mobile/src/main/java/net/ambitious/android/httprequesttile/sync/SyncActivity.kt
@@ -1,0 +1,14 @@
+package net.ambitious.android.httprequesttile.sync
+
+import android.app.Activity
+import android.os.Bundle
+import androidx.work.OneTimeWorkRequestBuilder
+import androidx.work.WorkManager
+
+class SyncActivity : Activity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    WorkManager.getInstance(this).enqueue(OneTimeWorkRequestBuilder<SyncWorker>().build())
+    finish()
+  }
+}
